refactor: Tuist 도입 및 프로젝트 전환 (Phase 6 - 1/5, 2/5)
- Tuist 4.178.0 설치 및 Project.swift, Tuist.swift, Package.swift 작성
- 타겟 2개(Mople, MopleDev) → 1개 + Debug/Release configuration 분리
- .xcodeproj를 Git 추적에서 제거 (tuist generate로 생성)
- Firebase xcframework 중복 문제 해결 (Analytics → ObjC 런타임 호출)
- 중복 파일 정리 (CommentListView 공백폴더, Comment.swift, DefaultMentionRepo 등)
- .gitignore에 Tuist 생성물 추가
