+++ Tuist.swift
@@ -1,0 +1,5 @@
+// Tuist 전역 설정 파일
+// tuist generate 시 프로젝트 생성 옵션을 정의한다
+import ProjectDescription
+
+let tuist = Tuist()
